fixing circular imports and adding another interface
